Working on ProfileTreeWidget
Thsi commit makes a bit more progress on the ProfileTreeWidget.  The
widget was changed to a generic QWidget with a private QTreeWidget
instead of inheriting from QTreeWidget, which seems like a much more
appropriate structure.  Items are also being created and inserted more
intelligently, and there is a basic context menu to expand and
collapse all items.