Update to profiler to handle unexpected case.
I was seeing data where the incremental inclusive cost of a node was
less than the sum of its childrens' costs.  It turns out that this was
because of the way we clamp time to one second.  This fix tries to
instead clamp the time to the actual duration since the last
publishing.